new version

+++ framework/v0.1.0/lib/generator/index.js
@@ -1,7 +1,7 @@
 /* Gina.Utils.Generator
  *
  * This file is part of the gina package.
- * Copyright (c) 2009-2014 Rhinostone <gina@rhinostone.com>
+ * Copyright (c) 2017 Rhinostone <gina@rhinostone.com>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -18,10 +18,8 @@ var fs = require('fs');
  * @namespace   Gina.Utils.Generator
  * @author      Rhinostone <gina@rhinostone.com>
  * */
-function Generator () {
-    var self  = {};
-
-    self.createFileFromTemplate = function(source, target, callback){
+var Generator = {
+    createFileFromTemplate : function(source, target, callback){
 
         fs.readFile(source, function(err, data){
             if (err) throw err;
@@ -29,28 +27,60 @@ function Generator () {
             if(fs.existsSync(target)){
                 //Just in case.
                 fs.chmodSync(target, 0755);
-                fs.unlink(target);
+
+                if ( fs.unlinkSync ) {
+                    fs.unlinkSync(target)
+                } else {
+                    fs.unlink(target);
+                }
             }
             fs.writeFile(target, data, function(err, data){
-                if (err) throw err;
-                //Setting permission.
-                fs.chmodSync(target, 0755);
-                if ( typeof(callback) != 'undefined') callback(err)
+                setTimeout( function onChmod(){
+                    if (err) throw err;
+                    //Setting permission.
+                    fs.chmodSync(target, 0755);
+                    if ( typeof(callback) != 'undefined') callback(err)
+                }, 1000)
             });
         });
-    }
+    },
+    /**
+     * Create file form template sync
+     * Added to support node v5.5.0
+     * */
+    createFileFromTemplateSync : function(source, target) {
+        var data = fs.readFileSync(source);
+        if (data instanceof Error) {
+            throw data
+        } else {
+            if(fs.existsSync(target)){
+                //Just in case.
+                fs.chmodSync(target, 0755);
 
-    self.createFileFromDataSync = function(data, target){
+                if ( fs.unlinkSync ) {
+                    fs.unlinkSync(target)
+                } else {
+                    fs.unlink(target);
+                }
+            }
+
+            fs.writeFileSync(target, data);
+
+            //Setting permission.
+            try {
+                fs.chmodSync(target, 0755)
+            } catch (err) {} // file not found
+        }
+    },
+    createFileFromDataSync : function(data, target){
         var data = (typeof(data) == "object") ? JSON.stringify(data, null, 4) : data;
         fs.writeFileSync(target, data);
         fs.chmodSync(target, 0755)
-    }
+    },
+    createFoldersFromStructureSync : function(structure){
 
-    self.createFoldersFromStructureSync = function(structure){
-
-    }
-
-    self.createPathSync = function(path, callback) {
+    },
+    createPathSync : function(path, callback) {
         var t = path.replace(/\\/g, '\/').split('/');
         var path = '';
         //creating folders
@@ -69,10 +99,7 @@ function Generator () {
         } catch (err) {
             callback(err);
         }
-    };
-
-    return self
+    }
 };
 
-
-module.exports = Generator()
+module.exports = Generator
